+++ src/customfieldlistdialog.h
@@ -53,7 +53,7 @@ private:
 
     wxDataViewListCtrl* fieldListBox_ = nullptr;
 
-    int m_field_id = -1;
+    int64 m_field_id = -1;
     mmCustomFieldListDialog() {}
     std::map<int, wxString> ColName_;
